+++ cgtjs/index.ts
@@ -1,6 +1,6 @@
 export * from './Board.ts';
 export * from './CanonicalForm.ts';
-export * from './game/Blokus.ts';
+export * from './game/Blokus/Blokus.ts';
 export * from './game/Konane.ts';
 export * from './MoveSet.ts';
 export * from './NumberUpStar.ts';
