+++ test/Blokus/MoveGeneration.test.ts
@@ -1,5 +1,5 @@
 import { describe, expect, test } from 'vitest';
-import { Blokus } from '../../cgtjs/game/Blokus';
+import { Blokus } from '../../cgtjs/game/Blokus/Blokus';
 import { uniqueBy } from '../../cgtjs/utils/unique';
 import { DOMINO, L4 } from './util/pieces';
